separate contact endpoints
